backend/requirements.txt
# Django & REST Framework
django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# LangChain & LLM
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.10
langchain-openai>=0.0.3
langchain-groq>=0.0.1
langgraph>=0.0.20
langchain-text-splitters>=0.0.1

# Vector Store
pinecone-client>=3.0.0

# LLM Providers
openai>=1.7.0

# PDF Processing
PyMuPDF>=1.23.8
pdf2image>=1.16.3
pytesseract>=0.3.10

# Research Tools
arxiv>=2.0.0
tavily-python>=0.3.0
dj-database-url>=2.1.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
scholarly

# PDF Export
reportlab>=4.0.7

# Utilities
pydantic>=2.5.3
python-dotenv>=1.0.0
requests>=2.31.0
